% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{blank_day}
\alias{blank_day}
\title{Creates a ITime vector spanning a single day with a custom time step}
\usage{
blank_day(timestep = NULL)
}
\arguments{
\item{timestemp}{The desired time step in minutes}
}
\description{
#' Creates a ITime vector spanning a single day with a custom time step
}
\examples{
blank_day()
}
\keyword{ITime}
