% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{blank_seq}
\alias{blank_seq}
\title{Create blank time sequence for each unique subject}
\usage{
blank_seq(from = NULL, to = NULL, subjects = NULL, timestep = NULL)
}
\arguments{
\item{from}{a start date as a character string ("yyyy-mm-dd") or an IDate object}

\item{to}{a start date as a character string ("yyyy-mm-dd") or an IDate object}

\item{subjects}{a character vector of subject IDs}

\item{timestep}{the desired time step in minutes}
}
\description{
Creates a data table with subject, date (IDate), and time (ITime) columns over
a specified range of days with a given time step
}
\examples{
blank_seq()
}

