% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{Circular}
\alias{Circular}
\title{Calculate the area of a circle}
\usage{
Circular(r)
}
\arguments{
\item{r}{Radius of the circle}
}
\description{
This function calculates the area of a circle given the radius
}
\examples{
Circular()
}
\keyword{Sap}
\keyword{flux}
