% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion_Functions.R
\name{conc_to_flux}
\alias{conc_to_flux}
\title{Convert change in gas concentration over time to flux}
\usage{
conc_to_flux(delta.gas.ppm, chamber.volume.cm3, collar.area.cm2, temperature.k,
  pressure.pa, species, input.time, output.units = "g/m2d1", C.only = FALSE,
  delta.ppm = NULL)
}
\arguments{
\item{delta.gas.ppm}{A vector of slopes representing the change in concentration 
over time}

\item{chamber.volume.cm3}{The gas chamber volume in cm^3; can be a vector of 
the same length as delta.ppm or a single numeric value}

\item{collar.area.cm2}{The area of the chamber footprint in cm^2; can be a vector
of the same length as delta.ppm or a single numeric value}

\item{temperature.k}{The temperature within the chamber in Kelvin; can be a 
vector of the same length as delta.ppm or a single numeric value}

\item{species}{The gas species of interest; 'CO2', 'CH4', or 'N2O'}

\item{input.time}{The time scale for delta.ppm; either 'seconds' or 'minutes'}

\item{C.only}{A logical specificying if the output flux should be calculated 
for the gas species or for carbon only; defaults to FALSE}

\item{delta.ppm}{Deprecated, use delta.gas.ppm}

\item{pressure.Pa}{The pressure within the chamber in Pascals;; can be a 
vector of the same length as delta.ppm or a single numeric value}

\item{ouput.units}{The desired output flux units; 'g/m2h1', 'mg/m2h1', 
'g/m2d1', or 'mg/m2d1'; defaults to 'g/m2h1'}
}
\value{
A vector giving flux of the gas species or of carbon in the selected 
  output units
}
\description{
This function uses the ideal gas law to convert changes in concentration 
within a sample chamber to flux rates for the gas species or for carbon. 
Conversion uses the ideal gas law to determine the convert the change in
concentration to the change in mass.
}

