% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Efflux_Gadget.R
\name{efflux}
\alias{efflux}
\title{Easily fit multilple linear regressions over selected data points}
\usage{
efflux(input.data, xvar, yvar, idvar)
}
\arguments{
\item{input.data}{A dataset containing a response variable and one or more
columns that can be used to create a unique ID}

\item{xvar}{The time step variable or any independent variable (quoted)}

\item{yvar}{Gas concentrations or any dependent variable (quoted)}

\item{idvar}{A single ID column (quoted) or a character vector containing the
names of the ID columns}
}
\description{
This shiny gadget provides an easy interactive method to trim sample data and
fit linear regressions to each sample's data. It is specifically designed for
soil and stem efflux. It will work with any type of data that has a unique
sample ID and dependent/independent variables.
}
\examples{
efflux()
}

