% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sap_Flux_Functions.R
\name{fluxmod}
\alias{fluxmod}
\title{Sap-flux response to Dz}
\usage{
fluxmod(a, w, x, z)
}
\arguments{
\item{a, w, x}{fitting paramters which need start values specified in nlmer.}

\item{z}{Dz}
}
\value{
The fitted values for a,w, and x for the equation \deqn{J[s]=a+w(1-x^{-D[z]})}
}
\description{
This function fits a curve to sap flux data using Dz as a predictor.  It is meant to be used within nlmer().
}
\examples{
fluxmod()
}
\keyword{Dz}
\keyword{flux}
\keyword{sap}

