% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{ggpanel}
\alias{ggpanel}
\title{Create a horizontal or vertical panel of plots with a shared legend}
\usage{
ggpanel(plots, orientation = "horizontal")
}
\arguments{
\item{plots}{A list of ggplot2 plots}

\item{orientation}{A character of "horizontal" or "vertical", defaults to 
'horizontal'.}
}
\value{
A ggplot object
}
\description{
Combines multiple plots generated by ggplot2 into a 1 column (vertical) or 1 
row (horizontal) panel with a shared legend.
}

