% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sap_Flux_Functions.R
\name{pataki_flow}
\alias{pataki_flow}
\title{Calculate sap flow for a tree with sapwood depth greater than probe length}
\usage{
pataki_flow(sap.flux, sapwood.area, diameter, tree.type)
}
\arguments{
\item{sap.flux}{Sap flux data as volume per unit area per unit time}

\item{sapwood.area}{Sapwood area of the tree in units matching
\code{sap.flux} area unit}

\item{diameter}{Diameter of the tree in units matching the square
root of the \code{sap.flux} area unit}

\item{tree.type}{is a character of 'Softwood' or 'Hardwood'}
}
\value{
Sap flow of the stem given as volume per unit time matching the units
  in \code{sap.flux}.
}
\description{
This function uses the equations laid out in Pataki et al, 2011 to calculate
the total sap flow of a tree from sap flux measurements in the outer sapwood.
It assumes a 2 cm long sap-flux probe and silently considers the total stem
area to be sapwood area if the sapwood area is greater than the area of the
tree caluclated form the diameter argument.
}
\examples{
pataki_flow()
}
\keyword{Sap}
\keyword{flux}
