% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sap_Flux_Functions.R
\name{pataki_var}
\alias{pataki_var}
\title{Calculate sap flow for a tree with sapwood depth greater than probe length}
\usage{
pataki_var(Flux, Var, Sapwood_Radius, Radius, Type)
}
\arguments{
\item{Flux}{Sap flux data}

\item{Var}{Variance of sap flux data}

\item{Sapwood_Radius}{Radius of the sapwood of the tree}

\item{Radius}{Radius of the tree at breast height (DBH/2)}

\item{Type}{Softwood or hardwood}
}
\description{
This function calculates the variance associated with the sap flow calculated using the equations laid out in Pataki et al, 2011.
}
\examples{
pataki_var()
}
\keyword{Sap}
\keyword{flux}
