% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{read_rdb}
\alias{read_rdb}
\title{Read in USGS rdb data}
\usage{
read_rdb(file)
}
\arguments{
\item{file}{A file in the USGS .rdb format}
}
\description{
This function is a wrapper for read.delim and was taken from the source of
\code{\link[dataRetrieval]{importRDB1}}. This function acts upon exisiting 
downloads and does not import data from a url
}
\examples{
read_rdb()
}
