% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{read_xle}
\alias{read_xle}
\alias{read_and_convert_xle}
\title{Read in Solinst .xle levellogger and barologger files}
\usage{
read_xle(file = NULL)

read_and_convert_xle(x, conversions = NULL)
}
\arguments{
\item{file}{A file in the .xle format}

\item{conversions}{A list of character vectors of length 2, each consisting of the input unit and the desired unit using abbreviations from \code{\link[measurements]{conv_unit}}}
}
\value{
A dataframe containing project ID, site, coordinates, date of measurement ('yyyymmdd'), time of measurement ('HH:MM:SS'), and the data extracted from each channel in the logger. If a channel includes measurement parameters such as offset or altitude these are included in the dataframe.
}
\description{
Use \code{read_xle} to read in Solinst data from levelloggers and barologgers that are in the *.xle format without exporting the file to a csv. Use \code{read_and_convert_xle} to convert the stored values as they are read in.
}
\examples{
read_and_convert_xle(file = "some_file.xle", 
                     conversions = list(c("ft", "m"), 
                                        c("F", "C")))
}
