% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{scientific_10x}
\alias{scientific_10x}
\title{Format numbers to scientific notation (10^#)}
\usage{
scientific_10x(values, digits = 1)
}
\arguments{
\item{values}{A numeric vector}

\item{digits}{A single integer value specifying the number of digits to
display after the decimal, trailing zeroes will be preserved}
}
\value{
An expression
}
\description{
Formats numbers to scientific notation using the form #.# x 10^# for display
on axes. Leading zeros and '+' signs are removed from the output and spaces
are inserted to match all exponents to uniform length for uniform display.
}
\examples{
x <- 1:3
y <- c(0.1, 100, 1000)

# Base Plotting
originalMargins <- par()$mar
plot(y ~ x, 
     axes = F, 
     par(mar = c(5, 5, 4, 2) + 0.1), ylab = NA)
axis(1)
axis(2, 
     at = y, 
     labels = scientific_10x(y), 
     las = 2)
par(mar = originalMargins)

# ggplot2
ggplot2::qplot(x, y) + 
  ggplot2::scale_y_continuous(labels = scientific_10x)
ggplot2::qplot(x, y) + 
  ggplot2::scale_y_continuous(breaks = y, 
                              labels = scientific_10x(y, digits = 2))

}
