% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{SEy}
\alias{SEy}
\title{Predicted Standard Error}
\usage{
SEy(model, Xes)
}
\arguments{
\item{model}{Any fitted model.}

\item{Xes}{the x-values used in prediction}
}
\description{
Calcualtes the standard error of predicted y-values
}
\examples{
SEy()
}
\keyword{error}
\keyword{standard}
