% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{ggpanel}
\alias{ggpanel}
\title{Create a horizontal or vertical panel of plots with a shared legend}
\usage{
ggpanel(plots, orientation = "horizontal", legend.position = "bottom")
}
\arguments{
\item{plots}{A list of ggplot2 plots}

\item{orientation}{A character of "horizontal" or "vertical", defaults to 
'horizontal'.}

\item{legend.position}{A character vector of "top", "bottom", "left", or "right", defaults to 'bottom'}
}
\value{
A ggplot object
}
\description{
Combines multiple plots generated by ggplot2 into a 1 column (vertical) or 1 
row (horizontal) panel with a shared legend.
}
\examples{
data1 <- data.frame(x = c(1:5, 1:5, 1:5, 1:5),treatment = rep(c(letters[1:2]), 
               each = 5, 
               times = 2),
year = rep(c("2015", "2016"), 
          each = 10),
power = c(rep(2, times = 5), 
         rep(2.1, times = 5), 
         rep(2.5, times = 5), 
         rep(3, times = 5))) \%>\% 
 mutate(y = rnorm(10, mean = 0, sd = 2) + x ^ power)


plot1 <- ggplot(data1, aes(x, y, color = treatment)) + geom_point() + geom_smooth(span = 1, se = F) + facet_grid(.~year) + theme_bw()
plot1


data2 <-
 data.frame(x = c(1:5, 1:5, 1:5, 1:5),
            treatment = rep(c(letters[1:2]), each = 5, times = 2),
            year = rep(c("2015", "2016"), each = 10),
            slope = c(rep(3, times = 5), rep(4, times = 5), rep(3, times = 5), rep(7, times = 5))) \%>\% 
 mutate(y = rnorm(20, mean = 0, sd = 2) + x * slope)
plot2 <- ggplot(data2, aes(x, y, color = treatment)) + geom_point() + geom_smooth#'(method = "lm", span = 1, se = F) + facet_grid(.~year) + theme_bw()
plot2

ggpanel(list(plot1, plot2), orientation = "vertical", legend.position = "right")
}

