% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sap_Flux_Functions.R
\name{pataki_flow}
\alias{pataki_flow}
\title{Calculate sap flow for a tree with sapwood depth greater than probe length}
\usage{
pataki_flow(Flux, Sapwood_Area, DBH, Type)
}
\arguments{
\item{Flux}{Sap flux data (m3m-2d-1)}

\item{Sapwood_Area}{Sapwood area of the tree (m2)}

\item{Diameter}{of the tree at breast height (m2)}

\item{Type:}{'Softwood' or 'Hardwood'}
}
\description{
This function uses the equations laid out in Pataki et al, 2011.
}
\examples{
pataki_flow()
}
\keyword{Sap}
\keyword{flux}

