% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{read_dir}
\alias{read_dir}
\title{Read in a directory of files}
\usage{
read_dir(dir = getwd(), pattern = NULL, collapse = TRUE, fun = read.csv,
  ...)
}
\arguments{
\item{dir}{A directory that contains your data files, defaults to the working directory}

\item{pattern}{A pattern to match filenames as in \code{\link[base]{list.files}}}

\item{collapse}{A logical argument, when true a single object is returned,
when false an object is returned for each file}

\item{fun}{The function to use to read in the files, defaults to 
\code{\link[base]{read.csv}}}

\item{...}{Additional arguments to pass to the input method}
}
\value{
When \code{collapse = T} a single object matching the output class
  of \code{fun} is returned. When \code{collapse = T} a list of objects
  matching the output class of \code{fun} is returned with names corresponded
  to the names of the imported files
}
\description{
This function reads in all files from a directory using the choosen import
function.  Use the 'pattern' argument to specificy a set of files, or a
single file type. If collapse = TRUE then all files must have identical
layouts.
}
\examples{
read_dir()
}

