% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{read_dir}
\alias{read_dir}
\title{Read in a directory of files with identical data formats}
\usage{
read_dir(dir = getwd(), pattern = NULL, fun = readr::read_csv, ...)
}
\arguments{
\item{dir}{A directory with only data files}

\item{pattern}{A pattern to match filenames as in \code{\link[base]{list.files}}}

\item{fun}{The function to use to read in the files, defaults to 
\code{\link[readr]{read_csv}}}

\item{...}{Additaional arguments to pass to the input method}
}
\description{
This function reads in all files from a directory.  All files must have the same layout.
}
\examples{
read_dir()
}

